Fix race condition in resource deployment...
The AWS template will occassionally fail to deploy with the error:

"The pending-instance-running instance to which 'eni-f00b4r'
is attached is not in a valid state for this operation (Service: AmazonEC2;
Status Code: 400; Error Code: IncorrectInstanceState; Request ID: FOOBAR)"

From looking into the issue, the public IP and network interface can
sometimes fail to associate with each other before the instance is up
and attempts to attach the interface. This commit adds a dependency to
all the VM instances on their corresponding IP association objects.

Ref - https://stackoverflow.com/questions/55389465/ec2-error-using-cloudformation-service-amazonec2-status-code-400-error-code