Fix issues with CentOS 8 image crashing after 1h
NetworkManager is essential in CentOS 8, disabling it causes a system
crash after 1 hour. Removing the disable/stop of NetworkManager fixes
the crashing issue for CentOS 8. This removal has been tested with a
CentOS 7 deployment (which definitely has NetworkManager running) and
allowed the system to work as expected. So no need to disable
NetworkManager at all.