Add Second Subnet for AWS (see commit body)
To prevent intermittent issues with network configuration, especially
between OS versions, the network for the compute nodes has been
separated and set to use the network interface of the chead node as the
gateway to the internet. This means that routing no longer needs to be
handled on a compute-node level but instead is carried out by the cloud
platform.

This change will also need to be applied to Azure.